medcor_backend2/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-spectacular==0.27.0
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
fastmcp==0.1.0
uvicorn==0.27.0
pydantic==2.5.3
bcrypt==4.1.2